pyproject.toml
[tool.poetry]
name = "eznashdb"
version = "0.1.0"
description = "The Ezrat Nashim Database"
authors = ["Ari Abramowitz <ari.abramowitz1@gmail.com>"]

[tool.pycln]
all = true
expand_stars = true

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 105

[tool.black]
line-length = 105
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | \.vscode
  | build
  | dist
)/
'''

[tool.djlint]
profile = "django"
blank_line_before_tag="block"
blank_line_after_tag="endblock,extends"
format_attribute_template_tags=true
max_line_length=105
preserve_leading_space=false
preserve_blank_lines=false

[tool.flakeheaven]
# PIE786 covered by bugbear with better message
ignore = ["E203", "E266", "E501", "W503", "F403", "F401", "E231", "-PIE786"]
format = "colored"
max_line_length = 105
max_complexity = 18
show_source = true
select = ["B","C","E","F","W","T4","B9"]
exclude = [".git","__pycache__","old","build","dist",".venv"]
baseline = "flakeheaven_baseline.txt"

[tool.flakeheaven.plugins]
pyflakes = ["+*"]
# E722 covered by bugbear with better message
pycodestyle = ["+*", "-E501", "-E722", "-W503"]
flake8-bugbear = ["+*"]
flake8-django = ["+*"]
flake8-pytest-style = ["+*"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
dj-database-url = "^0.5.0"
python-dotenv = "^0.19.2"
django-cors-headers = "^3.11.0"
psycopg2-binary = "^2.9.5"
black = "^23.3.0"
django-filter = "^23.2"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
pytest-django = "^4.5.2"
pytest-testmon = "^1.2.2"
pytest-watch = "^4.2.0"
pytest-spec = "^3.2.0"
beautifulsoup4 = "^4.11.1"
black = "^23.3.0"
pytest-xdist = "^3.2.1"
isort = "^5.12.0"
pre-commit = "^3.3.2"
djlint = "^1.19.7"
pycln = "^2.1.3"
flakeheaven = "^3.3.0"
flake8-pytest-style = "^1.7.2"
flake8-bugbear = "^22.1.11"
flake8-comprehensions = "^3.12.0"
flake8-django = "^1.2"
flake8-mutable = "^1.2.0"
pytest-describe = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
